requirements.txt
streamlit>=1.32.0
tensorflow>=2.10.0
numpy>=1.24.0
Pillow>=9.0.0
plotly>=5.15.0
